// src/app/models/conductor.ts
export class Conductor {
    condId?: Number;
    condTipoDocumento: String | undefined;
    condDocumento: String | undefined;
    condNombreCompleto: String | undefined;
    condDireccion: String | undefined;
    condCiudad: String | undefined;
    condDepartamento: String | undefined;
    condPais: String | undefined;
    condTelefono: String | undefined;
}